Remove unnecessary test
This test is using snapshot files (.li) that were generated once, but
is lacking a way to generate new ones. Furthermore the tests are only
checking that the correct settings are present in LayoutInspector.

The test don't seem very useful but are hard to maintain because of
the .li files.

Test: NA
Bug: NA
Change-Id: I93d109057481e8a6c109b331c24334dcfd087b64