Remove the VIEW_IN_FOCUS_MODE feature flag
View in Focus is a small and low risk functionality we added to Compose
Preview in change I5d033e7. By being overly cautious and adding it
behind a flag, initially set to "upToCanary" instead of true, we ended
up forgetting to enable the functionality in stable.

Instead of flipping the flag to COMPLETE, we decided to remove the flag
completely, considering this small improvement doesn't require a feature
flag.

Fixes: 438132171
Test: ViewInFocusModeActionTest
Change-Id: I2ee1ba30790fe0e2c9c5c10ce8d2b7f24f57138e