Convert last remaining AndroidGradleTestCase tests to JUnit4
This last set is full of weirdness, where the tests were written to
use not-complete (yet patched and plausible) projects, without
syncing, to see whether various operators can work on them.  Some of
these tests (and indeed the functionality that they are testing) are
of dubious value.

One difference between the JUnit3 and JUnit4 platform environment
manifested itself in these changes: it seems that the JUnit4 runner is
more likely to re-index asynchronously after a change to files, which
in particular affects running of refactoring processors (such as
AddRepoProcessor, underlying AddGoogleMavenRepositoryHyperlink).
Insert some waitForIndexes() calls to robustify the test.

Bug: none filed
Test: existing (test-only change)
Change-Id: I7ad91146cf62b4f54db3c5def43367e3c245e622