Include compile jars when generate dependency graph
export target's jar will not be listed in jdeps' dependencies. Use
compile_jars to include it. It's only necessary for first level
dependencies but the count of jars does not change a lot. Use photo
project as an example, include it for all targets will lead to ~80 more
jars. Consider photos have ~9k jars, it's not a big difference. So
inlcude compile_jars for every target.

Bug: b/439825455
Test: added
Change-Id: Ib81f9699837fcf56f8793e36173965f299a47b7e