Remove AGP and Gradle from android.core test_data
Historically, many AGP-related tests were in the android.core test
module.  This started being discouraged in 2021, with the addition of
a test to check that only allowlisted uses of gradle-related test
classes in android.core, and since about 2023 there have been no AGP-
or Gradle-related tests in android.core at all.

It is therefore now possible to remove AGP and Gradle from the tests'
available data, which means that any attempt to add a new AGP-related
or Gradle-related test to this target will fail.  That's probably
sufficient deterrent, so remove the single-use check_tests.bzl rule
and its ImlModuleCheckTests java backing implementation.

We can also remove the data dependency on
//tools/base/resource-repository by testing file extensions using an
apk that is already present in //tools/adt/idea/android/testData.

Bug: none filed
Test: existing (test-only change)
Change-Id: I37373efc6f4eb24cb8219e470b999ace683601bd