Use a fixed step session clock for static preview callbacks
When rendering a preview, different pieces of UI code may schedule
different callbacks to happen at different times in the future.

In this context, having a real time clock makes sense for non-static
preview, where constant renders are executed. But for static preview
it is better to use a fixed step clock, to ensure a deterministic
behavior.

Bug: 438554021
Bug: 431932822
Test: SteppingSessionClockTest
Change-Id: Ica5ebeb9d09299ab84bb0299aeeeff604142ffd0