Ease on GradleProjectPath
This is mainly motivated by one code path where we try to update the
build variant view even when sync fails (and this method ends up being
called). Without this, if any test is exercising deliberate failures
they are more prone to irrelevant failures. The other alternative would
be to remove this update behaviour as that also doesn't make much sense
on failure.

I also thought about whether we should make this nullable instead, but
almost all callers seem like they should be lenient to this type of
failure, and it's hard from the caller's perspective to figure out
whether no dynamic feature exists or this is a failure,
so it made more sense to make the implementation this way.

Bug: 384022658
Test: n/a
Change-Id: Ib75d7bb33a0f5e3622faf34133c910e5f67a371c