[AUA] Handle the compose compiler plugin transition
If we're upgrading through 9.0, we might need to modify
compose-using (Gradle) projects to apply the Compose Kotlin Compiler
Gradle plugin, rather than use composeOptions.  We detect this if a
project turns compose on but does not have the kotlin compose plugin
applied.

The change to 2.2.10 as the Kotlin version is perhaps not
necessary (yet) but is done to facilitate integration testing, because
we already have the artifacts for that version in prebuilts.

The behaviour of this processor on a project with multiple modules
with compose turned on was manually verified.

Bug: none filed
Test: added
Change-Id: I2e2f28a6d81d72653d3abd576593403b34cc0783