Fix reference tag references in quick edit fragments
When editing expressions in quick edit fragments, the references were
not being created as the module associated to literal expression was
null. Since the module was null, we were not able to check that the
watch face format version was 4 or above.

This CL fixes the issue by getting the module from the watch face file
instead.

Bug: b/443685010
Test: added regression test to ReferenceTagReferenceTest
Change-Id: Ic89775958b572945874929613a62e6b6c48a9fd0