Fix template parameter string reference variants in quick
... edit fragments

When autocompleting expressions in quick edit fragments, the string
references were not being computed properly. This is because the
StudioResourceRepositoryManager associated with the psiElement was null.

This CL fixes the issue by getting the android facet from the injection
host. This facet is then used to get the
StudioResourceRepositoryManager.

The android facet in the resolve method is not null when in a quick
edit fragment and did not need updating.

Fixes: b/443685010
Test: added regression test to TemplateParameterStringReferenceTest
Change-Id: Ib4517946e5be9c5230f9e0d626199f4505596be7