Add studiobot.knowledge.base.rag to StudioFlags
We want to use the flag for Knowledge Base (KB) RAG. The goal of KB is
adding the agent tools to support searching Android developer API docs
like docs for Jetpack Compose and other libraries.

This commit does not set a value for the flag, so it will be "false" as
the default value. We may need privacy review before dogfooding.

Bug: 441572280
Test: N/A
Change-Id: Id5cc4adb792b82969a3f6d78db7ee72368e29d08