If project names repeat - add paths for clarity
In Build settings we now show a list of projects affected
by Manual Sync and 'per project snooze'. If projects'
names aren't unique - we would show repeated names which
could look confusing. In that scenario we will now show
this in a form of list, with project paths in brackets
for extra clarity.

If project names are unique, the old way of showing of a
comma-separated list will still be used.

Bug: 438128341
Test: SyncDueMessageTest
Change-Id: I0e4455b5d0021b927c84fee52235c617b2a50180