Add junit5 classes to classpath for tests
After ag/34330969 IntelliJ finds junit5 classes on the classpath and
tries to use parts of it when constructing a multi-test test run i.e.
an entire directory or selecting test files. This results in:
"java.lang.NoClassDefFoundError: org/junit/platform/engine/TestDescriptor"
since some of these classes are not available on the classpath.

For now, we add the missing classes to the classpath. In the future we
will clean this up as part of b/439631949.

Fixes: 444018992
Bug: 439631949
Test: manual
Change-Id: I9cbe8f6e87fd4f0073489a53b094267e5c334428