Address misleading message in bundle generation
There are several reasons why it might not be possible to generate a
Bundle for the user's selection.  Now that we don't support AGP 3.2
projects, blaming it on too low a version of AGP is not reasonable.

Also, clean up the truly awful way the warning message was being
generated to show in the UI.

Bug: 216744700
Test: manual
Change-Id: I4b4c26f8d42478c0bc7a47a0abb1164e981694ea