Correctly render data in documentation popups upon hover
Text in the documentation popup upon hover would render with
incorrect line breaks. This was caused by an adjacent, empty `<td>`
with a `width="100%"` attribute, which was corrupting the layout for
the renderer.

This change introduces logic in `AndroidJavaDocExternalFilter` to find
and remove these specific empty `<td>` elements from the HTML before it
is rendered, resolving the text wrapping issue.

Bug: 442777828
Test: AndroidJavaDocExternalFilterTest
Change-Id: I8eabd2d85127a5531c74c00360f06ee707a1ca0b