Relax locking in AndroidProjectRule
Specifically, make sure that when opening a project we run on the EDT
(with the write intent lock) for as little as possible and as
interruptibly as possible, to allow potential background thread
writers to make progress.

Remove @RunsInEdt annotations from some withIntegrationTestRule()
users so that they don't force running on the EDT (which re-introduces
the deadlock); add runInEdtAndWait { } around sections which do
require write intent or write locks.  Explicitly wait for smart mode
in one place that otherwise exposes us to a potential delay from not
running everything on the EDT any more.

Also re-enable background thread writers.

Bug: 430533136
Test: existing
Change-Id: I778802b95a46303e338660c0a89bb8a3a0a5dec9