Improve class loader disposal
Before this change, the usage of a deprectaed isDisposed method
was causing us to never detect a disposed class loader as such,
and keep using it.

This change fixes the disposal detection and adapts the ViewLoader
to the now likely null result when loading a class.

This change is not a complete fix of the issue, but it significantly
reduces the memory leak.

Bug: 445923198
Bug: 443741623
Bug: 441253579
Bug: 445947658
Test: some adapted, other manually
Change-Id: I47b8e09776f5506fe8ca76879317421c9d01efb8