Moved more tests to use fromDir instead of fromZip
Updated all android studio end to end integration tests to use the optimized fromDir utility, so that the tests depend directly on the android studio directory instead of the zipped studio file.

as-driver/Build: Removed the dependency to the zipped studio directory.
studio/Build: Removed the manual tag for WebpSupportTest.
Added the dependency to the zipped studio file for UpdateTest since it will still continue to use the fromZip utility. UpdateTest modifies the Ide's resources jar and build number in the build.txt file
These operations are not possible with our fromDir utility, since the studio directory is read-only for all tests

Test: n/a
Bug: 444599236
Change-Id: Ia65c54acf59af259768e402b7f6709f8a353b8e7