Use diffPercent attribute for screenshot match %
This change refactors how the screenshot match percentage is
obtained. Previously, it was parsed from the error stack trace, which
was brittle.

The `diffPercent` attribute is now retrieved from the test plugin,
providing a more robust way to calculate the match percentage. This
also ensures that the match percentage is displayed for passing tests.

Bug: 441668746
Test: ScreenshotAttributesViewTest, ScreenshotTestUtilsTest. Manually tested with passing and failing screenshot tests.
Change-Id: I8edeef352067da010a7c9d10bc6fea469292513d