Update memoryAxis range in Live Telementry
This CL adds the MemoryAxis to the list of updatable components in
LiveMemoryFootprintModel to ensure it is refreshed along with the memory usage data.

This change fixes a bug where the Y-axis of the memory graph in the
"View Live Telemetry" profiler view was not updating, leading to an incorrect scale.

Bug: 377230194
Test: N/A
Change-Id: Iedc3a7eacb59fc04445502dea20f9e6a74fbd3ff