Add support to templates for configuring a Journeys 'test suite'
Templates can now be configured with a `TestSuiteWidget`. This widget
follows the same pattern as `PackageNameWidget` to bind the value
entered by the user to the `model.testSuiteName` property. This allows
the validator to access the chosen `testSuiteName` when checking if the
file names are unique. It also allows us to suggest defaults based on
context, i.e. if the user launches the template from within a test suite
then we can default the widget to the name of that test suite.

This CL also contains an update to the `DefaultRecipeExecutor` to
configure a Journeys test suite and tests for the new Journey template
which makes use of the `TestSuiteWidget`.

Bug: 433523178
Test: golden/testJourneysWithTestSuite
Change-Id: Icc4009af52eabce044820abfe9286da5a7747aee