Fix ResourceImportDialogViewModel validation message
The validation message in `ResourceImportDialogViewModel` was not being
updated correctly when the asset set was empty. This was because the
`doValidateAll()` method, which has side-effects of updating the error
message, was not being called when the asset set was empty.

This change ensures that `doValidateAll()` is always called, even when
the asset set is empty, so that the error message is always up-to-date.

Additionally, the format of the validation message has been updated to
be more consistent.

Fixes: 442763194
Test: ResourceImportDialogViewModelTest
Change-Id: I0f782b8d63daa74fbd4d6f441f15cb012720aca5