Replace emulator snapshot migration with removal
The `failureRetention` and `emulatorSnapshots` DSL blocks in
`testOptions` are no longer supported.

This patch updates the AGP Upgrade Assistant to reflect this. It removes
the previous
`MigrateFailureRetentionToEmulatorSnapshotsRefactoringProcessor`, which
migrated properties from one deprecated block to the other.

It introduces a new `RemoveEmulatorSnapshotsRefactoringProcessor` that
completely removes both the `failureRetention` and `emulatorSnapshots`
DSL blocks from build files.

All related tests, test data, and UI messages in
`AgpUpgradeBundle.properties` are updated to align with this new removal
behavior.

Bug: 328748781, 328748875, 328748995, 328749036
Test: updated
Change-Id: I95d20acaaf5d6468eaf47e9e5a417dd1b73e2b18