Allow ability to add images in the prompt
This is the first pass for adding images to the prompt. There's a decent amount
of hardcoding for now since a lot of these are taken from aiplugin.

We may want to have a better data structure for when we want to show the images
in the prompt itself (instead of the counter).

Test: TBD
Fixes: 437277653
Change-Id: If54ff6d17d89fd1605948055452e500528048b8c