Improve image factory logic in RenderTask
Before this change, when quality was 1f, an unbounded image was
being created, which is dangerous. This change makes sure that all
images created by the factory in RenderTask are bounded to a limit.

Before this change, there was a bug in the constraints calculations.
Quality and max size constrains were being multiplied sometimes, now
the tighter one of them is used.

Before this change, different device types would get different image
size limits. This logic hasn't been updated for a long time, and there
are new big device types that are not being considered as special
cases. Also sometimes a custom device defined by a user may be some
custom deskptop or similar device, which should then be treated as
a desktop. This change addresses these issues by using a unique max
image size limit.

Bug: 425743229
Test: ConstrainedImageFactoryTest
Change-Id: I3097ab58855f9ea44d1d81a2c2b97d2829a3ea1f