Remove the toValues lambda from findAnnotatedMethodsValues
... and rename the method to findAnnotatedMethods.

The toValues calculation is now the responsibility of the caller.

The toValues parameter was preventing caching from working in cases
where the toValues was a lambda. This is because the lambda is different
on each call, meaning the key was different on each call too. This was
not caught by the tests, as the tests were passing functions which have
the same identity on each call.

The toValues was computed within findAnnotatedMethodsValues under a read
action. This CL adds readActions where needed as the toValues has been
moved out of  findAnnotatedMethodsValues.

Bug: 381827960
Bug: 397205975
Test: existing
Change-Id: Ifcf32ebe25dc268e5ddef8c6da2fbe74d90e6988