Allow user to switch to wifi pairing code tab if QR pairing fails
Currently if QR pairing fails, and user switches to pairing code tab,
the user won't see any pairing services since the polling has stopped.

This bug exists in the current v1 dialog. The fix only targets the v2
dialog. This CL is a follow up to the comment in ag/33533130.

Test: manual
Bug: 412571872
Change-Id: I1968b206805dad86fc831573afe162753bc9c3b8