Fix IndexNotReady exception
The IndexNotReady was happening in
MultiRepresentationPreview#updateRepresentationsImpl. This was because
ComposePreviewRepresentationProvider#accept was returning true, even in
DumbMode. In isCompatibleComposableClassAvailable we were checking if we
were in dumb mode or not, however that check was only done if psiFile
.getModuleSystem()?.usesCompose was false.

All other preview providers return false in their accept method when in
dumb mode. This is to prevent calling PreviewRepresentation#hasPreviews
which was also causing IndexNotReady. By not accepting in dumb mode, the
hasPreviews method is not called.

Fixes: b/398265392
Test: added regression test to ComposePreviewRepresentationProviderTest
Change-Id: Ie65603c70e8fb59029e2a68dfd7ea97aa78ec611