Remove calls to AndroidDebugBridge.hasInitialDeviceList
Remove unnecessary calls to `AndroidDebugBridge.hasInitialDeviceList()`
from production code. This check is redundant because we immediately
iterate over `AndroidDebugBridge.devices`. If the initial device list
has not yet been processed, the device list will simply be empty, which
is handled correctly.

`AndroidDebugBridge.hasInitialDeviceList()` is still used in some tests
that rely on polling.

Bug: n/a
Test: existing
Change-Id: I5befeca08237effc980ec500087b4485e2152e91