Show agent plan in the thinker
Implement the TODO plan in the thinker, both in the expanded and
collapsed states. There are a bunch of animations throughout, which are
just an attempt and can be tweaked as needed.

The expanded state has a max height, and will be scrollable if the plan
doesn't fit in it. The current item will always be visible when the
thinker is expanded, but the user can freely scroll around.

The feature is behind an always-off flag. We also add a flag for tl;dr
mode (aka exec mode) for later use.

Bug: none
Test: none
Change-Id: Ie703f12ef5c73be0375cd4eb8d03d85323c00b92