Update permission-versions.xml output format in auto-complete PPD test.
There were some naming changes suggested for the permission-versions.xml
output format. Updating the tests to reflect these naming changes. Note
that these attributes have no impact on the actual auto-complete PPD
logic as versioning info is ignored. So, the update in the test file is
mostly just for correctness.

Bug: 442571488
Test: AndroidManifestPurposeDomTest
Change-Id: Ic7d063e6f0faa3da8812998c8d09fed9e6ea0a76