Refactor background handling in PreviewDisplaySettings
Use a sealed class `Background` instead of separate `showBackground` and `backgroundColor` properties. This provides a more structured way to represent the different background options: `None`, `Default`, and `Color`.
This models better the options available to set the background in the
preview without allowing for cases that do not make sense like
`showBackground = false` with `backgroundColor = 0xFF0000`.

Test: Existing tests updated and pass.
Bug: N/A
Change-Id: I41041ef2166a2e971f85c45cf6e2fbeb89770d8b