Include source target with generated class jar
Previously we only include source targets that contains generated source
jar/ files. It leads to the annotation processor generated jars get
ignored. With this cl, we will attached the generated class jar of
source target. For a in project target, its JavaArtifacts.jars is only
contains the generated class jar from java_output, so it should not
contains the compile_jars that we do not want to include.

Bug: b/448400351
Test: added
Change-Id: I0ff22f4d868c2983d8a59fee46581b2309fa4b93