Remove TODO for b/424551554 as Module resource invalidation is architecturally correct.
Removes the TODO comments that requested replacement of the ModuleOutOfBlockModificationEvent fix.

The resource invalidation fix was investigated by the Kotlin team, who confirmed that module-level invalidation is the correct and most granular tool currently available for changes affecting R.string across the module. The temporary fix is now considered the permanent and intended solution.

Bug: 424551554
Test: N/A
Change-Id: I190450bf9e24402ef026154557fe9d1cc4964c07