[AUA] Add meta-checks in tests
It would be easy when editing version number lists to accidentally end
up with the same version number twice.  This would be a mistake in the
test generation, so detect it and cause the test to fail.  Likewise,
an incomplete editing operation might cause the version list to become
unordered; enforce ordering on test initialization.

Bug: none filed
Test: existing
Change-Id: Ife986b8ad184a927d407720096035179aace6ac3