Drop override for SDK custom processing flag
With IntelliJ 2025.1.1 custom processing for global libraries and SDKs
in WorkspaceFileIndex was disabled by default. However, in studio the
old behaviour was re-enabled by using the registry flag:
ide.workspace.model.sdk.remove.custom.processing, since change was
causing several test failures linked to Android SDK not being properly
indexed after being added to the JDK table.

The root issue, is the fact that SDK is instantiated when project
object was already created, but the proejct state is not yet in open
state. This issue is something IntelliJ faced as well, and recommended
fix is to create the SDK either before project object is created or
when the project is already in open state.

For this reason, the following changes has been made:

(1) GradleProjectImporter set up the SDK before init before project
object instance was created to align with new processing behaviour
for libraries and SDKs.

(2) LightJavaCodeInsightFixtureAdtTestCase doesn't require to remove
the SDK added by LightJavaCodeInsightFixtureTestCase since this is
controlled by the parent class, missing SDK references since doesn't
get indexed again.

Note: Existing flag will be removed in IntelliJ 2025.3.

Bug: 416578927
Test: existing, e.g. SpecificActivityLocatorGradleTest
Change-Id: Ifa64b6313d7b3a357ceb3575a8c409957b682ddc