Fix BuildProjectBleakTest
There's a new field in ProjectNotificationsModel which holds
notifications; make sure we ignore references from it.

Also, the class is ProjectNotificationsModel, and not
ProjectNotificationModel.

Finally, since these notifications in any case hold references to the
project, make sure that we expire old notifications when we are about
to create new ones so that we have a constant number of references.
This isn't strictly speaking a leak but pacifying the leak checker is
probably easier than informing the leak checker about the semantics of
project notifications.

Bug: 445656379
Test: this is it
Change-Id: Ib6833d88fc5e4abb00bc489aafc9d569f425320a