Refactor: Centralize QuerySync dispatcher logic
Moves the dispatcher creation and configuration to a new
QuerySyncDispatchers object. This allows a shared dispatcher to be used
across different query sync operations, such as in
JavaPackagePrefixReaderImpl and ParallelPackageReader.

The number of threads for the IO dispatcher is configurable via the
"querysync.reader.threads" experiment.

This change improves code organization and allows for easier management
of threading resources within QuerySync.

Bug: 448702414
Test: NA
Change-Id: I73a7dd13fe988eeba00559af158dec25889e9e71