Fix the tests filter logic for ss test classes
The Gradle run configuration for a single screenshot test class was
generating a `--tests` filter with a trailing wildcard. This could lead
to unexpected behavior by running tests in other classes that shared the
same name prefix.

The `--tests` filter for a single class should not use a wildcard.
This ensures only the selected class's tests are run, preventing
unintended execution of tests from other classes with same name prefix.

Bug: 450652533
Test: Existing. Tested manually.
Change-Id: Ibe7f9c5a2538a2526df7be317f22cef496160f7b