Move the zoom masks logic into ZoomController
Lot's of bugs where related with the race condition between
DesignSurface and the zoom-to-fit that has being notified too early
before the DesignSurface had the right size.

We adopted this logic only for Preview but we are now extending to all
the NlDesignSurfaces.
This will also fix issues with surfaces showing big zoom on drawables

Fixes: 445109082
Fixes: 374663154
Fixes: 425997179
Test: DesignSurfaceSettingsTest, VisualizationFormTest, ComposePreviewRepresentationGradleTest
Change-Id: I2805bb93775c3a9597ce53be4e5a411108a7fe09