Add analytics for LeakCanary task
This change introduces analytics tracking for the completion of the LeakCanary profiler task.
This will allow us to gather metrics on feature usage for both newly recorded sessions and for loading past recordings.
The following TaskFinished events are now tracked:
1. When a user-initiated LeakCanary recording is stopped.
2. When a previously recorded LeakCanary session is successfully loaded.
The start of a LeakCanary recording is already tracked via the common TaskEntered event,
which applies to all profiler tasks.
StudioFeatureTracker has been updated to recognize the LEAKCANARY task type.

Bug: 451516707
Test: Manually verified
Change-Id: Icbf38bb5f4afd9f2797557ae0e5b4b9f316bd415