Get selected render issue from DataContext
Refactor the `FixRenderIssuesAction` and related providers to retrieve
the selected issue from the `DataContext` using
`PlatformDataKeys.SELECTED_ITEM`. Previously, the issue was passed
explicitly through constructors and method arguments.

This change simplifies the creation and invocation of the action, as the
calling code no longer needs to explicitly pass the issue object. It
also unifies how the action retrieves information from the context,
aligning it with how other data, such as the virtual file, is accessed.

Bug: 444223371
Test: FixRenderIssuesActionTest
Test: DesignerCommonIssueSidePanelTest
Test: SceneViewErrorsPanelTest
Change-Id: Ie0573ba7e2ff9b39c6074fab8f9904161701a3de