Show warning in Wi-Fi Device Manager entries if device not up-to-date
adbd with the wifi v2 changes publish an mdns service with "2.0" in the
mdnsServiceVersion.

Old devices may not be eligible for an update, we can't detect that so
the warning message won't be actionable for such devices. It should be
ok to show the message since it doesn't stop the pairing process.

Bug: 412571872
Test: WifiPairableDeviceProvisionerPluginTest
Change-Id: I7fea2929405df39624fb0bb3d2110246c83e3928