Fix wireless pairing generic QR code scanning problem
Before this change, if user clicks on pairing action to open pairing
dialog for device A, then the user picks up device B and scans the QR
code with device B, the pairing succeeded for device B.

With new changes to adbd and adb host, we now have the serial of the
device in the pairing mdns service. We use the serial to make sure that
the device scanning the QR code is the target device.

This change also aligns QR code with pairing code since pairing code
only accepts the target device to be paired.

Test: manual
Bug: 412571872

Change-Id: Id5dc4c16d253f29189756262628560f006d39da3