Profiler: Log event on profiler window creation
This change adds a new telemetry that tracks when the profiler window is first clicked.

Rationale:
The goal of this event is to measure the discoverability of the Android Studio Profiler UI.
Tracking an event every time the profiler window is shown will add excessive logging.
By logging only upon the profiler window's initial creation, we capture the true point of user discovery of Profilers.

Bug: 454247833
Test: N/A
Change-Id: I7b4c8c808907d359994d561ebe8a6b2d21ce8e13