Add PreviewAnnotationProvider service
This commit introduces the `PreviewAnnotationProvider` project service.
This service calculates and provides a flow of all FQNs in the project
that are considered Preview annotations. This includes the base
`androidx.compose.ui.tooling.preview.Preview` annotation, and any
annotation classes or type aliases that are directly or transitively
annotated with or aliased to a Preview annotation.

The service uses Kotlin Analysis API to perform the resolution and
identification of annotations and type aliases. It is backed by a
StateFlow that is automatically updated when relevant code or module
structure changes occur. The calculation is debounced and cancellable.
The service also optimizes by only performing the full calculation if
the base `@Preview` annotation exists in the project dependencies.

Test: PreviewAnnotationProviderTest.kt
Fixes: 454570647
Change-Id: I2862aac16baeb4313ad8bf131a0cf79f4758bbc2