Prepare for setting up cc sources from graph only
and the rest of the configuration from the artifact tracker state.

Split the setup into two phases: (1) that uses build graph data only
and (2) that uses the artifact tracker state only.

This results in CcTargets being partially configured from the build
graph even when they are not yet available int he artifact tracker.

Bug: n/a
Test: existing
Change-Id: Iacb505358f1af2f9381ad643b595ae1a8d6f8a2f