Include proto src jar field
It's a new field only for java_proto_library/java_lite_proto_library. We
need it because java_proto_library did not include a source jar of its
generated jar field in java_info.java_outputs. It's listed in srcjar
field. But we only want to include srcjar of such target not everyone,
so we need a separate function to fetch its from these targets and set
them in a separate field, so that we can access and know which srcjars
should be used.

Bug: b/447640817
Test: added
Change-Id: I4f1b036a14a267b8f86b03196bc247d99f867f88