Add PreviewDefinitionProvider project service
This service provides a hot flow of all resolved Compose `@Preview` and
multi-preview definitions in the project. It handles changes in Kotlin
files and module roots to keep the list up-to-date. The service uses the
Kotlin Analysis API to resolve annotations and extract relevant
information. It also includes caching for multi-preview expansions to
improve performance and handle recursive definitions.

Test: PreviewDefinitionProviderTest.kt
Fixes: 455843969
Change-Id: I52ec210026553f256ccce08e64449ef4806cbfb4