AndroidPlatformSdk without project view set
Make it possible to obtains AndroidSdkPlatform from an skd string and
an minSdk value without providing a full project view set.

The implementation may still internally request the location of the
top level project view file to report it to the user. This is a leaky
abstraction that allows breaking the dependency on the query sync
project on the project view set, which makes it difficult to find out
which values from the project view set are consumed by the query sync.

Bug: n/a
Test: existing
Change-Id: Ib2103284623a379e5818803a6be4b34be916bed7