Introduce QuerySyncLanguageSettings
and replace direct usage of ProjectViewSet and similar generic
configuration containers in order to make it transparent what query sync
depends on.

An upcoming change will remove ProjectViewSet usage from the query sync
internals.

Considering that Android support is implemented in an separate target
that depends on the query sync, query sync targets cannot depend on
the Android implementation although the query sync implementation
hardcodes support for Android. Work it around by inverting the
dependency with the help of QuerySyncLanguageSettingsProvider.

Bug: n/a
Test: existing
Change-Id: I0da549ed11f71ec71b7692a32d1b8057097c6750